Some of PR comments
