+++ ocf_data_sampler/select/spatial_slice_for_dataset.py
