+++ ocf_data_sampler/torch_datasets/process_and_combine.py
@@ -141,3 +141,13 @@ def fill_nans_in_arrays(batch: dict) -> dict:
             fill_nans_in_arrays(v)
 
     return batch
+
+
+def compute(xarray_dict: dict) -> dict:
+    """Eagerly load a nested dictionary of xarray DataArrays"""
+    for k, v in xarray_dict.items():
+        if isinstance(v, dict):
+            xarray_dict[k] = compute(v)
+        else:
+            xarray_dict[k] = v.compute(scheduler="single-threaded")
+    return xarray_dict
